Use array as data type for languagePreference
The definition states that the learner's language preference is a "comma-seperated list". That would require an implementation to parse the property from a string into an array to process it properly over and over again. 

So the spec could use an array as data-structure directly. (Note: The JSON specification defines an array as "an ordered collection of values")